feat: add minor ui changes
