+++ Payan/App/Domain Modules/Hero/UI/View Model/PYHeroViewModel.swift
@@ -11,7 +11,6 @@ class PYHeroViewModel: ObservableObject {
     @Published var hero: PYHero = .empty
     @Published var isLoading = true
     @Published var errorHasOccured = false
-    @Published var isImageViewerVisible = false
     
     private var heroWasFetchedSuccessfuly = false
     private let interactor: PYHeroBusinessLogic
@@ -25,10 +24,6 @@ class PYHeroViewModel: ObservableObject {
             return " "
         }
         return "\(hero.bornAt) - \(hero.diedAt)"
-    }
-    
-    func showImageViewer() {
-        isImageViewerVisible = true
     }
     
     func section(at index: Int) -> PYHeroSection {
